# doc/api.yml
openapi: 3.0.3
info:
  version: "latest"
  title: Blogify - API
  description: |
    Log in and blog on [Blogify-web](https://github.com/blogify-app/blogify-web) 
    to blog and interact with other users using [Blogify-api](https://github.com/blogify-app/blogify-api) 
    as your giveaway provider. Comment, create posts, follow new trends and react to posts as you see fit.
security:
  - BearerAuth: []
paths:
...
  "/users":
    get:
      tags:
        - User
      summary: Get all users.
      operationId: getUsers
      parameters:
        - name: page
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Page"
        - name: page_size
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/PageSize"
        - name: name
          in: query
          required: false
          description: Filter user by first name or lastname or username.
          schema:
            type: string
      responses:
        "200":
          description: List of users that meet the criterias.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  "/users/{id}":
    get:
      tags:
        - User
      summary: Get user by identifier.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      operationId: getUserById
      responses:
        "200":
          description: The identified user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
    put:
      tags:
        - User
      summary: Crupdate user by identifier.
      operationId: crupdateUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: User to crupdate.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: The crupdated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  "/posts/{pid}/comments/{cid}/reaction":
    post:
      tags:
        - Comments
      summary: React to a comment by identifier.
      operationId: reactToCommentById
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
        - name: cid
          in: path
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/ReactionType"
      responses:
        "201":
          description: The comment reaction.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reaction"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"

  ## Commenting resources
  "/posts/{pid}/comments":
    get:
      tags:
        - Comments
      summary: Get the identified post comments.
      operationId: getCommentsByPostId
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Page"
        - name: page_size
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/PageSize"
      responses:
        "200":
          description: The identified post comments.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  "/posts/{pid}/comments/{cid}":
    put:
      tags:
        - Comments
      summary: Crupdate comment of the identified post by its identifier.
      operationId: crupdateCommentById
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
        - name: cid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: The comment to crupdate.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Comment"
      responses:
        "200":
          description: The crupdated comment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
    get:
      tags:
        - Comments
      summary: Get comment by identifier.
      operationId: getCommentById
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
        - name: cid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The identified comment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
    delete:
      tags:
        - Comments
      summary: Delete comment by identifier.
      operationId: deleteCommentById
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
        - name: cid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The deleted comment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  responses:
...
    "400":
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BadRequestException"
    "403":
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NotAuthorizedException"
    "404":
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ResourceNotFoundException"
    "429":
      description: Too many requests to the API
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TooManyRequestsException"
    "500":
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/InternalServerException"
  schemas:
...
    Page:
      type: integer
      description: Start at `1`
    PageSize:
      type: integer
      description: Cannot exceed `500`
...
    Sex:
      type: string
      enum:
        - M
        - F
        - OTHER
    ReactionType:
      type: string
      enum:
        - LIKE
        - DISLIKE
    UserStatus:
      type: string
      enum:
        - ENABLED
        - BANISHED
...
    Category:
      properties:
        id:
          type: string
        label:
          type: string
...
    CommentStatus:
      type: string
      enum:
        - ENABLED
        - DISABLED
    Comment:
      properties:
        id:
          type: string
        user:
          $ref: "#/components/schemas/User"
        post_id:
          type: string
        content:
          type: string
        creation_datetime:
          type: string
          format: date-time
        reply_to_id:
          type: string
        reactions:
          $ref: "#/components/schemas/ReactionStat"
        status:
          $ref: "#/components/schemas/CommentStatus"
    ReactionStat:
      properties:
        likes:
          type: number
        dislikes:
          type: number
    Reaction:
      properties:
        id:
          type: string
        type:
          $ref: "#/components/schemas/ReactionType"
        creation_datetime:
          type: string
          format: date-time
        user_id:
          type: string
        post_id:
          type: string
        comment_id:
          type: string
    User:
      properties:
        id:
          type: string
        last_name:
          type: string
        first_name:
          type: string
        birth_date:
          type: string
          format: date
        email:
          type: string
        photo_url:
          type: string
        bio:
          type: string
        profile_banner_url:
          type: string
        username:
          type: string
        about:
          type: string
        status:
          $ref: "#/components/schemas/UserStatus"
        sex:
          $ref: "#/components/schemas/Sex"
        entrance_datetime:
          type: string
          format: date-time
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        is_followed:
          type: boolean
...
    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    BadRequestException:
      allOf:
        - $ref: "#/components/schemas/Exception"
      example:
        type: BadRequestException
        message: Bad request
    NotAuthorizedException:
      allOf:
        - $ref: "#/components/schemas/Exception"
      example:
        type: NotAuthorizedException
        message: Not authorized
    ResourceNotFoundException:
      allOf:
        - $ref: "#/components/schemas/Exception"
      example:
        type: ResourceNotFoundException
        message: Resource of type <T> identified by <I> not found
    TooManyRequestsException:
      allOf:
        - $ref: "#/components/schemas/Exception"
      example:
        type: TooManyRequestsException
        message: Too many requests
    InternalServerException:
      allOf:
        - $ref: "#/components/schemas/Exception"
      example:
        type: InternalServerException
        message: Unexpected error